# Output json optimizations