Mostly provoked by this [community forum post](https://community.k6.io/t/k6-is-memory-hungry-when-using-several-modules/3000/33). 

Moving to using easyjson gave some improvement, but then I noticed we are spending a bunch of time in compression so decided to try out a different gzip implementation.

Combined improvement is twice better on both a CPU and allocations
```
name            old time/op    new time/op    delta
FlushMetrics-8    37.2ms ± 6%    18.2ms ± 6%  -50.99%  (p=0.000 n=10+9)

name            old alloc/op   new alloc/op   delta
FlushMetrics-8    2.44MB ± 1%    2.52MB ± 0%   +3.28%  (p=0.000 n=10+10)

name            old allocs/op  new allocs/op  delta
FlushMetrics-8     47.3k ± 1%     20.1k ± 0%  -57.59%  (p=0.000 n=10+10)
```

Breaking change: the tag keys are now *not* sorted in the output. This is due to no longer using the stdlib marshalling which does this.