@MStoykov 2022-03-11
Some of those can likely be used in the output/cloud as well as they don't seem that complicated to apply