@MStoykov 2022-03-16
Note: if we change the time to be int64 instead of `time.Time` we can get even better performance
Compared to the optimized version here:
```
name            old time/op    new time/op    delta
FlushMetrics-8    18.2ms ± 6%    11.3ms ±12%  -37.96%  (p=0.000 n=9+10)

name            old alloc/op   new alloc/op   delta
FlushMetrics-8    2.52MB ± 0%    1.85MB ± 0%  -26.64%  (p=0.000 n=10+9)

name            old allocs/op  new allocs/op  delta
FlushMetrics-8     20.1k ± 0%      7.6k ± 0%  -62.31%  (p=0.000 n=10+10)
```

Compared to the original:
```
name            old time/op    new time/op    delta
FlushMetrics-8    37.2ms ± 6%    11.3ms ±12%  -69.59%  (p=0.000 n=10+10)

name            old alloc/op   new alloc/op   delta
FlushMetrics-8    2.44MB ± 1%    1.85MB ± 0%  -24.23%  (p=0.000 n=10+9)

name            old allocs/op  new allocs/op  delta
FlushMetrics-8     47.3k ± 1%      7.6k ± 0%  -84.01%  (p=0.000 n=10+10)
```

Unfortunately I can't figure out a not terrible way to support both so we can be backwards compatible :( 

Also, of note is that I remembered there is a slight breaking change - the tag keys are now not sorted in the JSON output.